split stream into multiple files
