+++ src/stream/mod.rs
@@ -27,7 +27,6 @@ pub use repeat::{repeat, Repeat};
 pub use stream::{Stream, Take};
 
 mod empty;
-mod min_by;
 mod once;
 mod repeat;
 mod stream;
